offering delete

+++ src/offer/offer.controller.ts
@@ -1,5 +1,4 @@
 import { NextFunction, Request, Response, Router } from "express";
-import ISession from "interfaces/session.interface";
 import { Schema, Types } from "mongoose";
 
 import HttpException from "../exceptions/Http.exception";
@@ -7,8 +6,10 @@ import IdNotValidException from "../exceptions/IdNotValid.exception";
 import OfferNotFoundException from "../exceptions/OfferNotFount.exception";
 import IController from "../interfaces/controller.interface";
 import IRequestWithUser from "../interfaces/requestWithUser.interface";
+import ISession from "../interfaces/session.interface";
 import authMiddleware from "../middleware/auth.middleware";
 import validationMiddleware from "../middleware/validation.middleware";
+import orderModel from "../order/order.model";
 import CreateOfferDto from "./offering.dto";
 import IOffer from "./offering.interface";
 import offerModel from "./offering.model";
@@ -17,6 +18,7 @@ export default class OfferController implements IController {
     public path = "/offers";
     public router = Router();
     private offer = offerModel;
+    private order = orderModel;
 
     constructor() {
         this.initializeRoutes();
@@ -103,12 +105,16 @@ export default class OfferController implements IController {
         }
     };
 
-    // LINK ./Offer.controller.yml#deleteOrder
+    // LINK ./Offer.controller.yml#deleteOffer
     // ANCHOR[id=deleteOffer]
     private deleteOffer = async (req: Request, res: Response, next: NextFunction) => {
         try {
             const id = req.params.id;
             if (Types.ObjectId.isValid(id)) {
+                const isOfferHasReference = await this.order.findOne({ details: { $elemMatch: { offer_id: id } } });
+                if (isOfferHasReference){
+                    next(new Refe(""))
+                }
                 const offer = await this.offer.findOne({ _id: id });
                 if (offer) {
                     await this.offer.findByIdAndDelete(id);
